requirements.txt
asgiref==3.8.1
autoflake==2.3.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.2
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.9
Faker==37.1.0
flake8==7.2.0
flake8-isort==6.1.2
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
sslcommerz_python_api==1.0.1
tzdata==2025.2
urllib3==2.4.0
